requirements.txt
argon2-cffi @ file:///C:/ci/argon2-cffi_1613037959010/work
asgiref==3.5.2
async-generator @ file:///home/ktietz/src/ci/async_generator_1611927993394/work
attrs @ file:///tmp/build/80754af9/attrs_1620827162558/work
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
backports.zoneinfo==0.2.1
beautifulsoup4==4.10.0
bleach @ file:///tmp/build/80754af9/bleach_1628110601003/work
Bottleneck==1.3.2
bs4==0.0.1
cachetools==5.0.0
Cartopy @ file:///C:/ci/cartopy_1636040271244/work
certifi==2021.10.8
cffi @ file:///C:/ci/cffi_1636542074879/work
charset-normalizer==2.0.11
click==8.0.4
colorama @ file:///tmp/build/80754af9/colorama_1607707115595/work
configparser==5.2.0
crayons==0.4.0
cryptography==36.0.1
cx-Oracle==8.3.0
cycler @ file:///tmp/build/80754af9/cycler_1637851556182/work
debugpy @ file:///C:/ci/debugpy_1637073815078/work
decorator @ file:///tmp/build/80754af9/decorator_1632776554403/work
defusedxml @ file:///tmp/build/80754af9/defusedxml_1615228127516/work
distlib==0.3.4
Django==4.0.4
eli5==0.11.0
entrypoints==0.3
et-xmlfile==1.1.0
filelock==3.6.0
Flask==2.1.2
fonttools==4.25.0
future==0.18.2
geojson==2.5.0
google-api-core==2.5.0
google-api-python-client==2.37.0
google-auth==2.6.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.54.0
graphviz==0.19.1
greenlet==1.1.2
h11==0.13.0
html5lib==1.1
httplib2==0.20.4
idna==3.3
imbalanced-learn==0.9.0
importlib-metadata @ file:///C:/ci/importlib-metadata_1638543108096/work
ipykernel @ file:///C:/ci/ipykernel_1633545574250/work/dist/ipykernel-6.4.1-py3-none-any.whl
ipython @ file:///C:/ci/ipython_1635944310712/work
ipython-genutils @ file:///tmp/build/80754af9/ipython_genutils_1606773439826/work
ipywidgets @ file:///tmp/build/80754af9/ipywidgets_1634143127070/work
itsdangerous==2.1.2
jedi @ file:///C:/ci/jedi_1611315808330/work
Jinja2 @ file:///tmp/build/80754af9/jinja2_1635780242639/work
joblib==1.1.0
JPype1 @ file:///D:/bld/jpype1_1636588042576/work
jsonschema @ file:///Users/ktietz/demo/mc3/conda-bld/jsonschema_1630511932244/work
jupyter==1.0.0
jupyter-client @ file:///tmp/build/80754af9/jupyter_client_1637148478538/work
jupyter-console @ file:///tmp/build/80754af9/jupyter_console_1616615302928/work
jupyter-core @ file:///C:/ci/jupyter_core_1636537213213/work
jupyterlab-pygments @ file:///tmp/build/80754af9/jupyterlab_pygments_1601490720602/work
jupyterlab-widgets @ file:///tmp/build/80754af9/jupyterlab_widgets_1609884341231/work
kiwisolver @ file:///C:/ci/kiwisolver_1612282606037/work
konlpy==0.6.0
KoreaNewsCrawler==1.50
lightgbm==3.3.2
lxml==4.8.0
MarkupSafe @ file:///C:/ci/markupsafe_1621528314575/work
matplotlib @ file:///C:/ci/matplotlib-suite_1638289827560/work
matplotlib-inline @ file:///tmp/build/80754af9/matplotlib-inline_1628242447089/work
mistune==0.8.4
mkl-fft==1.3.1
mkl-random @ file:///C:/ci/mkl_random_1626186184278/work
mkl-service==2.4.0
munkres==1.1.4
nbclient @ file:///tmp/build/80754af9/nbclient_1614364831625/work
nbconvert @ file:///C:/ci/nbconvert_1624479163777/work
nbformat @ file:///tmp/build/80754af9/nbformat_1617383369282/work
nest-asyncio @ file:///tmp/build/80754af9/nest-asyncio_1613680548246/work
nltk==3.7
notebook @ file:///C:/ci/notebook_1637143794296/work
numexpr @ file:///C:/ci/numexpr_1640704408721/work
numpy @ file:///C:/ci/numpy_and_numpy_base_1634106873763/work
oauth2client==4.1.3
oauthlib==3.2.0
olefile @ file:///Users/ktietz/demo/mc3/conda-bld/olefile_1629805411829/work
opencv-python==4.5.5.64
openpyxl==3.0.9
outcome==1.1.0
packaging @ file:///tmp/build/80754af9/packaging_1637314298585/work
pafy==0.5.5
pandas @ file:///C:/ci/pandas_1635488579061/work
pandocfilters @ file:///C:/ci/pandocfilters_1605102497129/work
parso @ file:///tmp/build/80754af9/parso_1617223946239/work
patsy==0.5.2
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow==8.4.0
platformdirs==2.5.2
prometheus-client @ file:///tmp/build/80754af9/prometheus_client_1637050397234/work
prompt-toolkit @ file:///tmp/build/80754af9/prompt-toolkit_1633440160888/work
protobuf==3.19.4
py-hanspell @ git+https://github.com/ssut/py-hanspell.git@8e993cf46f97f9d665c15633a0fc78ac1b727713
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
Pygments @ file:///tmp/build/80754af9/pygments_1629234116488/work
PyMySQL==1.0.2
pyOpenSSL==22.0.0
pyowm==3.2.0
pyparsing @ file:///tmp/build/80754af9/pyparsing_1635766073266/work
pyperclip==1.8.2
pyrsistent @ file:///C:/ci/pyrsistent_1636111468851/work
pyshp @ file:///tmp/build/80754af9/pyshp_1610641829514/work
PySocks==1.7.1
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
python-twitter==3.5
pytz==2021.3
pywin32==302
pywinpty==0.5.7
pyzmq @ file:///C:/ci/pyzmq_1638435185959/work
qtconsole @ file:///tmp/build/80754af9/qtconsole_1632739723211/work
QtPy @ file:///tmp/build/80754af9/qtpy_1629397026935/work
regex==2022.3.15
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-learn==1.0.2
scipy @ file:///C:/ci/scipy_1630606942584/work
seaborn @ file:///tmp/build/80754af9/seaborn_1629307859561/work
selenium==4.1.0
Send2Trash @ file:///tmp/build/80754af9/send2trash_1632406701022/work
Shapely==1.7.1
sip==4.19.13
six @ file:///tmp/build/80754af9/six_1623709665295/work
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
SQLAlchemy==1.4.31
sqlparse==0.4.2
statsmodels==0.13.2
tabulate==0.8.9
terminado==0.9.4
testpath @ file:///tmp/build/80754af9/testpath_1624638946665/work
threadpoolctl==3.0.0
tornado @ file:///C:/ci/tornado_1606942392901/work
tqdm==4.62.3
traitlets @ file:///tmp/build/80754af9/traitlets_1636710298902/work
trio==0.19.0
trio-websocket==0.9.2
tweepy==4.5.0
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.8
virtualenv==20.14.1
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
webdriver-manager==3.5.3
webencodings==0.5.1
Werkzeug==2.1.2
widgetsnbextension==3.5.1
wincertstore==0.2
wordcloud==1.8.1
wsproto==1.0.0
xgboost==1.5.2
youtube-dl==2021.12.17
zipp @ file:///tmp/build/80754af9/zipp_1633618647012/work
